buffer.h: introduce spi related definitions
spi transactions will be described by a specific buffer operation with type
SPIS or SPIM (for slave or master). In such case the address value shall
contain the size in bits of the command, address, mosi and miso phases.
Should tx and rx be required within the same transaction, an sglist with two
buffers (one for tx and one for rx ) is expected as operand.