drivers, dev: allow the device side to force pipe closing
If a buffer is pushed on a pipe with NULL pointer and 0 length, the
pipe's circular buffer is reset, the read high watermark is set to
0xffffffff and an input ready event for the pipe itself is issued.
This way, the application will try reading and get zero bytes (as opposed
to -EAGAIN in case nothing is in the circular buffer). As a consequence
the pipe will be closed.