lib/tcp, client connection: setup connection data before calling tcp_connect().
When implementing a server, we get a new tcp_conn_data structure as soon as a
new connection has been established. On the other hand, a client needs a
connection to be allocated before tcp_connect() is called. That's needed
because if the active connection fails to be established, the error callback
must be provided with a pointer to a tcp_conn_data structure.

If the active connection cannot
be established,